+++ ailerons_tracker_backend/csv_parser/csv_parser.py
@@ -33,7 +33,7 @@ def parse_csv(path, csv_id: str):
         and a list of matching individual_ids """
     try:
         absolute_path = os.path.abspath(path)
-        df = pd.read_csv(
+        record_df = pd.read_csv(
             absolute_path,
             index_col=None,
             encoding='ISO-8859-1',
@@ -41,12 +41,11 @@ def parse_csv(path, csv_id: str):
             on_bad_lines='error',
             sep=';')
 
-        df_list = []
+        record_list = []
         individual_id_list = supabase.get_individual_ids()
         new_individual_id_list = []
 
-        for row in df.itertuples(index=False):
-
+        for row in record_df.itertuples(index=False):
             if not any(entry['individual_id'] == row.individual_id for entry in individual_id_list):
                 individual_id_list.append({'individual_id': row.individual_id})
                 new_individual_id_list.append(
@@ -54,9 +53,9 @@ def parse_csv(path, csv_id: str):
 
             new_record = Record(row._asdict())
             new_record.csv_id = csv_id
-            df_list.append(new_record.__dict__)
+            record_list.append(new_record.__dict__)
 
-        return df_list, new_individual_id_list
+        return record_list, new_individual_id_list
 
     except Exception as e:
         raise ParserError() from e
