+++ ailerons_tracker_backend/models/individual_model.py
