--- tracker-requests/bruno.json
@@ -1,5 +1,0 @@
-{
-  "version": "1",
-  "name": "tracker-requests",
-  "type": "collection"
-}
